credentials_service/backend/requirements.txt
flask==3.0.3
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
pytest==8.3.3
gunicorn==23.0.0
fasteners==0.19
